backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
pillow==10.1.0
numpy==1.24.3
accelerate==0.24.1
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2